backend/pyproject.toml
[tool.poetry]
name = "birdcall"
version = "0.1.0"
description = "api to feed with data from xenocanto"
authors = ["claydsonassis <claydsonassis@gmail.com>"]
license = "mit"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
mysql-connector-python = "^8.0.32"
sqlalchemy = "^1.4.46"
psycopg2 = "^2.9.5"
python-dotenv = "^0.21.1"
fastapi-users = "^10.3.0"
fastapi-users-db-sqlalchemy = "^4.0.4"
tqdm = "^4.64.1"
requests = "^2.28.2"


[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
